Fix issue when running original qunit task
When running original qunit task but with this plugin loaded in grunt,
grunt will use our qunit.js & phantom.js files.

So we need to verify that the environment variable for TAP results output
is set in the current process before trying to save the TAP results files.

+++ tasks/qunit/phantom.js
@@ -41,16 +41,18 @@ var resultsFilename = suiteName
 		.replace(/^\-|\-$/g, "")
 		.replace(/\-+/g, "-");
 
-var buildOutpuDir = system.env['QUNIT_TAP_OUTPUT'];
+var buildOutputDir = system.env['QUNIT_TAP_OUTPUT'];
 
-if (!fs.exists(buildOutpuDir)) {
-	fs.makeDirectory(buildOutpuDir);
-}
+if (buildOutputDir) {
+	if (!fs.exists(buildOutputDir)) {
+		fs.makeDirectory(buildOutputDir);
+	}
 
-resultsFilename = buildOutpuDir + resultsFilename + ".tap";
+	resultsFilename = buildOutputDir + resultsFilename + ".tap";
 
-if (fs.exists(resultsFilename)) {
-	fs.remove(resultsFilename);
+	if (fs.exists(resultsFilename)) {
+		fs.remove(resultsFilename);
+	}
 }
 
 // Messages are sent to the parent by appending them to the tempfile.
@@ -61,7 +63,9 @@ function sendMessage(args) {
 
 	// Exit when all done.
 	if (/^done/.test(args[0])) {
-		fs.write(resultsFilename, args[5], "a");
+		if (buildOutputDir) {
+			fs.write(resultsFilename, args[5], "a");
+		}
 		phantom.exit();
 	}
 }
